使用高性能同步等待

+++ src/isolate/per_isolate_data.h
@@ -23,7 +23,7 @@ class PerIsolateData final {
 
   static PerIsolateData* From(v8::Isolate* isolate) {
     return static_cast<PerIsolateData*>(
-        isolate->GetData(IsolateData::kPerIsolateData));
+        isolate->GetData(static_cast<uint32_t>(IsolateData::kPerIsolateData)));
   }
 
   V8TemplateMap& GetV8TemplateMap() { return template_map_; }
